separate fiat and sfincs config from hazard and impacts configs. make dark mode work, update styles for docs
