--- docs/_static/dark.scss
@@ -1,88 +1,0 @@
-/*-- scss:defaults --*/
-/*-- scss:rules --*/
-
-:root {
-  --background: #212529;
-  --highlight: #373a3c;
-  --highlight-section: #D3D3D3;
-  --link-color: rgb(62, 200, 255);
-  --text-color: #ffffff;
-
-  --callout-bg: #2a2f35;
-  --callout-border: #3a3f47;
-  --callout-text: #e0e0e0;
-  --callout-header-text: #2d2c2c;
-
-  --doc-section-bg: #3a3f47;
-  --code-bg: #2e3238;
-  --code-text: #e0e0e0;
-}
-
-/* General styling */
-body {
-  background-color: var(--background);
-  color: var(--text-color);
-}
-
-h1, h2 {
-  color: var(--link-color);
-}
-
-/* Callouts */
-.callout {
-  background-color: var(--callout-bg);
-  border: 1px solid var(--callout-border);
-  color: var(--callout-text);
-}
-
-.callout-title {
-  color: var(--callout-header-text);
-}
-
-/* Other components */
-.content {
-  margin-left: 100px;
-}
-
-.nav-page-text {
-  margin-left: 80px;
-}
-
-table {
-  table-layout: auto !important;
-  width: 100% !important;
-}
-
-th, td {
-  white-space: nowrap;
-}
-
-.doc-section dt code {
-  background: none;
-}
-
-.doc-section dt {
-  background-color: var(--doc-section-bg);
-  display: none;
-}
-
-.doc-section dl dd {
-  margin-left: 3rem;
-}
-
-.navbar-right .fa-github {
-  font-size: 40px;
-}
-
-/* Theme-specific visibility */
-.only-on-dark {
-  display: none;
-}
-
-.only-on-light {
-  display: none;
-}
-
-/*-- scss:mixins --*/
-/*-- scss:functions --*/
-/*-- scss:uses --*/
