+++ docs/_static/images/logo.png
