+++ docs/_static/styles/light.scss
