+++ docs/scripts/clean_database.py
