+++ docs/scripts/version.py
@@ -1,5 +1,5 @@
 from pathlib import Path
 from flood_adapt import __version__
 
-with open(Path(__file__).parent / "_version.yml", "w") as f:
+with open(Path(__file__).parent.parent / "_version.yml", "w") as f:
     f.write(f"version: {__version__}")
