+++ flood_adapt/adapter/sfincs_adapter.py
@@ -931,7 +931,11 @@ class SfincsAdapter(IHazardAdapter):
                 timeseries=tmp_path, magnitude=None, direction=None
             )
         elif isinstance(wind, WindMeteo):
-            ds = MeteoHandler().read(time_frame)
+            ds = MeteoHandler(
+                dir=self.database.static_path / "meteo",
+                lat=self.database.site.lat,
+                lon=self.database.site.lon,
+            ).read(time_frame)
             # data already in metric units so no conversion needed
 
             # HydroMT function: set wind forcing from grid
@@ -1015,7 +1019,11 @@ class SfincsAdapter(IHazardAdapter):
 
             self._model.setup_precip_forcing(timeseries=tmp_path)
         elif isinstance(rainfall, RainfallMeteo):
-            ds = MeteoHandler().read(time_frame)
+            ds = MeteoHandler(
+                dir=self.database.static_path / "meteo",
+                lat=self.database.site.lat,
+                lon=self.database.site.lon,
+            ).read(time_frame)
             # MeteoHandler always return metric so no conversion needed
             self._model.setup_precip_forcing_from_grid(precip=ds, aggregate=False)
         elif isinstance(rainfall, RainfallTrack):
