+++ flood_adapt/config/gui.py
@@ -7,7 +7,7 @@ import pandas as pd
 import tomli
 from pydantic import BaseModel, Field, model_validator
 
-from flood_adapt.config.fiat import DamageType
+from flood_adapt.config.impacts import DamageType
 from flood_adapt.objects.forcing import unit_system as us
